[Refactor]: refactoring user schema
