+++ app/schemas/response/character_response_schema.py
@@ -2,7 +2,6 @@ from enum import Enum
 from fastapi_camelcase import CamelModel
 
 
-
 class CharacterGenderEnum(str, Enum):
     male = 'male'
     female = 'female'
@@ -16,3 +15,6 @@ class CharacterResponse(CamelModel):
     character_personality: str | None = None
     character_details: str | None = None
     user_id: int
+
+class MessageResponse(CamelModel):
+    message: str
