+++ tests/test_user.py
@@ -48,7 +48,7 @@ def test_read_users(db_session: Session):
     assert response.status_code == 200
     data = response.json()
     assert len(data) > 0
-    assert data[0]["user_email"] == "test@example.com"
+    assert data[0]["userEmail"] == "test@example.com"
 
 def test_read_current_user(db_session: Session):
     """
@@ -64,8 +64,8 @@ def test_read_current_user(db_session: Session):
     response = client.get(f"/api/user/me", headers=headers)
     assert response.status_code == 200
     data = response.json()
-    assert data["user_name"] == "Test User"
-    assert data["user_email"] == "test@example.com"
+    assert data["userName"] == "Test User"
+    assert data["userEmail"] == "test@example.com"
 
 def test_update_current_user(db_session: Session):
     """
@@ -80,12 +80,12 @@ def test_update_current_user(db_session: Session):
 
     response = client.put(
         f"/api/user/me",
-        json={"user_name": "Updated User"},
+        json={"userName": "Updated User"},
         headers=headers
     )
     assert response.status_code == 200
     data = response.json()
-    assert data["user_name"] == "Updated User"
+    assert data["message"] == "User updated successfully"
 
 def test_deactivate_user(db_session: Session):
     """
